Seprarte viewState from ViewUtils
viewState is used as a slim down model for a specific TL, viewUtil contains alot of render helper functions to construct or view. As our views get more interactive and complex, it longer makes sense to put them in the same file. As it prevents us from providing viewState componenet models, because of circular depedency issues