Do ast change actions on model update tick
JC: is it possible to accidentally make an AST update without going through your update function? If it is possible, which I suspect it might be,
I think it might be preferable to do something similar to the virtual DOM: at the end of every model tick, diff the current ast with the ast from the previous frame. If they differ, run analysis. If they don't differ, don't run analysis.

A: We can do it, though the newAST <> oldAST is the way we've been diffing all over our app. I am not sure if this is going to be slow as AST grows bigger and may greatly slow our client down